Refactor token routes and rename related files and imports.
Renamed `tokens_routes.py` to `git_tokens.py` for clarity and updated file imports and usages accordingly. Removed unused code related to adding tokens and Git hosting provider integration, simplifying the token route logic.
