+++ app/routes/__init__.py
@@ -4,10 +4,10 @@ Routes module initialization.
 
 from fastapi import APIRouter
 
-from app.routes.tokens_routes import router as tokens_routes
+from app.routes.git_tokens import router as git_tokens
 
 # Create main router
 router = APIRouter()
 
 # Include sub-routers
-router.include_router(tokens_routes, prefix="/tokens", tags=[""])
+router.include_router(git_tokens, prefix="/git_tokens", tags=[""])
